aws-scripts/s3-incomplete-mpu/s3-mpu-cdk/requirements.txt
aws-cdk-lib==2.18.0
constructs>=10.0.0,<11.0.0
black==22.3.0
